default duration for images, cleanup storage
