+++ frontend/package.json
@@ -15,33 +15,33 @@
   },
   "dependencies": {
     "@nuxtjs/color-mode": "^3.5.2",
-    "@pinia/nuxt": "^0.9.0",
+    "@pinia/nuxt": "^0.10.1",
     "@vueform/multiselect": "^2.6.11",
     "@vuepic/vue-datepicker": "^11.0.1",
-    "@vueuse/nuxt": "^12.5.0",
+    "@vueuse/nuxt": "^12.7.0",
     "bootstrap-icons": "^1.11.3",
     "dayjs": "^1.11.13",
     "jwt-decode": "^4.0.0",
     "lodash-es": "^4.17.21",
     "mpegts.js": "^1.8.0",
     "nuxt": "3.15.4",
-    "pinia": "^2.3.1",
+    "pinia": "^3.0.1",
     "sortablejs-vue3": "^1.2.11",
     "splitpanes": "^3.1.8",
     "video.js": "^8.21.0"
   },
   "devDependencies": {
-    "@nuxt/eslint": "^0.7.5",
-    "@nuxtjs/i18n": "^9.1.1",
+    "@nuxt/eslint": "^1.1.0",
+    "@nuxtjs/i18n": "^9.2.1",
     "@nuxtjs/tailwindcss": "^6.13.1",
     "@types/lodash-es": "^4.17.12",
     "@types/video.js": "^7.3.58",
     "daisyui": "^4.12.23",
     "mini-svg-data-uri": "^1.4.4",
-    "postcss": "^8.5.1",
+    "postcss": "^8.5.3",
     "postcss-loader": "^8.1.1",
-    "sass": "^1.83.4",
-    "sass-loader": "^16.0.4",
+    "sass": "^1.85.0",
+    "sass-loader": "^16.0.5",
     "vue": "^3.5.13",
     "vue-router": "^4.5.0"
   }
